\name{spacetravel-package}
\alias{spacetravel-package}
\alias{spacetravel}
\docType{package}
\title{\packageTitle{spacetravel}}
\description{\packageDescription{spacetravel}}
\section{Package Content}{\packageIndices{spacetravel}}
\author{\packageAuthor{spacetravel}}
\section{Maintainer}{\packageMaintainer{spacetravel}}
\keyword{package}
